pyproject.toml
[tool.poetry]
name = "invoice-manager"
version = "0.1.14"
description = ""
authors = ["Wioletta Wajda <wioletta.wajda@devsoft.pl>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.11"
celery = "^5.2.7"
django-environ = "0.11.2"
djangorestframework = "^3.14.0"
black = "24.4.0"
isort = "^5.12.0"
django-filter = "24.2"
django-extensions = "^3.2.1"
xhtml2pdf = "^0.2.9"
num2words = "^0.5.12"
pytest-django = "^4.5.2"
coverage = "^7.2.5"
factory-boy = "^3.2.1"
requests = "2.32.2"
redis = "5.0.3"
parameterized = "^0.9.0"
django-factory-boy-generator = "^1.0.4"
django-cors-headers = "^4.3.1"
django-storages = "^1.14.2"
gunicorn = "22.0.0"
psycopg = "^3.1.18"
jinja2 = "3.1.4"
sentry-sdk = "^2.3.1"
flake8 = "^7.0.0"
boto3 = "^1.34.103"
uvicorn = "^0.29.0"

[tool.poetry.group.semver.dependencies]
python-semantic-release = "^9.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''

(
  /(
    | \.git
    | migrations
    | venv
    | \.env
  )/
)
'''

[tool.isort]
skip = [".git", "migrations", "venv", ".env"]
profile = "black"

[tool.semantic_release]
version_toml = ["pyproject.toml:tool.poetry.version"]
commit_message = "chore(release): {version}"
tag_format = "{version}"

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease = false
prerelease_token = "beta"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["build", "chore", "ci", "docs", "feat", "fix", "perf", "style", "refactor", "revert"]
minor_tags = ["feat"]
patch_tags = ["build", "chore", "ci", "fix", "perf", "style", "refactor", "revert"]


[tool.semantic_release.changelog]
template_dir = "semantic_template"
changelog_file = "CHANGELOG.md"
exclude_commit_patterns = ["build", "chore", "ci", "docs", "perf", "refactor", "revert", "style"]

[tool.semantic_release.remote]
name = "origin"
type = "github"
